+++ distmaker/dists/joomla_php5.sh
@@ -37,10 +37,12 @@ cp "$SRC/joomla/admin/admin.civicrm.php" "$DM_TMPDIR/com_civicrm/admin/civicrm.p
 cd $DM_TMPDIR;
 
 # generate alt version of package
-cp -R -p civicrm com_civicrm/admin/civicrm
-${DM_PHP:-php} $DM_SOURCEDIR/distmaker/utils/joomlaxml.php $DM_SOURCEDIR com_civicrm $DM_VERSION alt
-${DM_ZIP:-zip} -q -r -9 $DM_TARGETDIR/civicrm-$DM_VERSION-joomla-alt.zip com_civicrm
-rm -rf com_civicrm/admin/civicrm
+if [ -z "$DM_SKIP_ALT" ]; then
+  cp -R -p civicrm com_civicrm/admin/civicrm
+  ${DM_PHP:-php} $DM_SOURCEDIR/distmaker/utils/joomlaxml.php $DM_SOURCEDIR com_civicrm $DM_VERSION alt
+  ${DM_ZIP:-zip} -q -r -9 $DM_TARGETDIR/civicrm-$DM_VERSION-joomla-alt.zip com_civicrm
+  rm -rf com_civicrm/admin/civicrm
+fi
 
 # generate zip version of civicrm.xml
 ${DM_PHP:-php} $DM_SOURCEDIR/distmaker/utils/joomlaxml.php $DM_SOURCEDIR com_civicrm $DM_VERSION zip
